requirements.txt.txt
pathway
llm-app
unstructured[all-docs]==0.10.15
tiktoken
requests
streamlit
python-dotenv